+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added form dateInput component
 - Added yarn audit check to dangerfile (on package change)
 - Added search component
+- Added Grid, GridContainer components
 
 ## [1.2.0] - 2020-04-30